removed the hard limit of x days no play, instead it now depends on the rating of a song, in the following manner:
if rating is the rating of the song with 0 < rating < 1 and bdays is
the number of days a song with 0.5 would not be played again:

block_days = rating**(log(bdays, 0.5))

also made some other simplifications because of this: no longer need
relaxors, just play a song more often or manually rate it higher to
hear it more often. Rate it lower/skip it to hear it less
often. Feedback loops ftw.